+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/DocumentAnalysisClient.java
@@ -3,11 +3,11 @@
 
 package com.azure.ai.formrecognizer;
 
-import com.azure.ai.formrecognizer.models.AnalyzeDocumentOptions;
-import com.azure.ai.formrecognizer.models.DocumentAnalysisException;
 import com.azure.ai.formrecognizer.implementation.models.AnalyzeResultOperation;
 import com.azure.ai.formrecognizer.implementation.models.OperationStatus;
+import com.azure.ai.formrecognizer.models.AnalyzeDocumentOptions;
 import com.azure.ai.formrecognizer.models.AnalyzeResult;
+import com.azure.ai.formrecognizer.models.DocumentModelOperationException;
 import com.azure.ai.formrecognizer.models.DocumentOperationResult;
 import com.azure.core.annotation.ReturnType;
 import com.azure.core.annotation.ServiceClient;
@@ -26,8 +26,16 @@ import static com.azure.ai.formrecognizer.implementation.util.Utility.toFluxByte
  * Operations allowed by the client are analyzing information from documents and images using custom-built document
  * analysis models, prebuilt models for invoices, receipts, identity documents and business cards, and the layout model.
  *
- * <p><strong>Instantiating a synchronous Document Analysis Client</strong></p>
- * {@codesnippet com.azure.ai.formrecognizer.DocumentAnalysisClient.instantiation}
+ * <p><strong>Instantiating an asynchronous Document Analysis Client</strong></p>
+ *
+ * <!-- src_embed com.azure.ai.formrecognizer.DocumentAnalysisClient.instantiation -->
+ * <pre>
+ * DocumentAnalysisClient documentAnalysisClient = new DocumentAnalysisClientBuilder&#40;&#41;
+ *     .credential&#40;new AzureKeyCredential&#40;&quot;&#123;key&#125;&quot;&#41;&#41;
+ *     .endpoint&#40;&quot;&#123;endpoint&#125;&quot;&#41;
+ *     .buildClient&#40;&#41;;
+ * </pre>
+ * <!-- end com.azure.ai.formrecognizer.DocumentAnalysisClient.instantiation -->
  *
  * @see DocumentAnalysisClientBuilder
  */
@@ -52,15 +60,31 @@ public final class DocumentAnalysisClient {
      * indicating absence of cancellation support</p>
      *
      * <p><strong>Code sample</strong></p>
-     * {@codesnippet com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string}
+     * <p> Analyze a document using the URL of the document. </p>
+     * <!-- src_embed com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string -->
+     * <pre>
+     * String documentUrl = &quot;&#123;document_url&#125;&quot;;
+     * String modelId = &quot;&#123;custom_trained_model_id&#125;&quot;;
      *
-     * @param modelId The unique model ID to be used or the supported prebuilt models - "prebuilt-receipt",
-     * "prebuilt-businessCard", "prebuilt-idDocument", "prebuilt-document", "prebuilt-invoice", "prebuilt-layout".
+     * documentAnalysisClient.beginAnalyzeDocumentFromUrl&#40;modelId, documentUrl&#41;.getFinalResult&#40;&#41;
+     *     .getDocuments&#40;&#41;.stream&#40;&#41;
+     *     .map&#40;AnalyzedDocument::getFields&#41;
+     *     .forEach&#40;documentFieldMap -&gt; documentFieldMap.forEach&#40;&#40;key, documentField&#41; -&gt; &#123;
+     *         System.out.printf&#40;&quot;Field text: %s%n&quot;, key&#41;;
+     *         System.out.printf&#40;&quot;Field value data content: %s%n&quot;, documentField.getContent&#40;&#41;&#41;;
+     *         System.out.printf&#40;&quot;Confidence score: %.2f%n&quot;, documentField.getConfidence&#40;&#41;&#41;;
+     *     &#125;&#41;&#41;;
+     *
+     * </pre>
+     * <!-- end com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string -->
+     *
+     * @param modelId The unique model ID to be used. Use this to specify the custom model ID or prebuilt model ID.
+     * Prebuilt model IDs supported can be found <a href="https://aka.ms/azsdk/formrecognizer/models">here</a>
      * @param documentUrl The URL of the document to analyze.
      *
      * @return A {@link SyncPoller} to poll the progress of the analyze document operation until it has completed,
      * has failed, or has been cancelled. The completed operation returns an {@link AnalyzeResult}.
-     * @throws DocumentAnalysisException If analyze operation fails and the {@link AnalyzeResultOperation} returns
+     * @throws DocumentModelOperationException If analyze operation fails and the {@link AnalyzeResultOperation} returns
      * with an {@link OperationStatus#FAILED}..
      * @throws IllegalArgumentException If {@code documentUrl} or {@code modelId} is null.
      */
@@ -77,10 +101,26 @@ public final class DocumentAnalysisClient {
      * error message indicating absence of cancellation support</p>
      *
      * <p><strong>Code sample</strong></p>
-     * {@codesnippet com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string-AnalyzeDocumentOptions-Context}
+     * <p> Analyze a document using the URL of the document with configurable options. </p>
+     * <!-- src_embed com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string -->
+     * <pre>
+     * String documentUrl = &quot;&#123;document_url&#125;&quot;;
+     * String modelId = &quot;&#123;custom_trained_model_id&#125;&quot;;
      *
-     * @param modelId The unique model ID to be used or the supported prebuilt models - "prebuilt-receipt",
-     * "prebuilt-businessCard", "prebuilt-idDocument", "prebuilt-document", "prebuilt-invoice", "prebuilt-layout".
+     * documentAnalysisClient.beginAnalyzeDocumentFromUrl&#40;modelId, documentUrl&#41;.getFinalResult&#40;&#41;
+     *     .getDocuments&#40;&#41;.stream&#40;&#41;
+     *     .map&#40;AnalyzedDocument::getFields&#41;
+     *     .forEach&#40;documentFieldMap -&gt; documentFieldMap.forEach&#40;&#40;key, documentField&#41; -&gt; &#123;
+     *         System.out.printf&#40;&quot;Field text: %s%n&quot;, key&#41;;
+     *         System.out.printf&#40;&quot;Field value data content: %s%n&quot;, documentField.getContent&#40;&#41;&#41;;
+     *         System.out.printf&#40;&quot;Confidence score: %.2f%n&quot;, documentField.getConfidence&#40;&#41;&#41;;
+     *     &#125;&#41;&#41;;
+     *
+     * </pre>
+     * <!-- end com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocumentFromUrl#string-string -->
+     *
+     * @param modelId The unique model ID to be used.  Use this to specify the custom model ID or prebuilt model ID.
+     * Prebuilt model IDs supported can be found <a href="https://aka.ms/azsdk/formrecognizer/models">here</a>
      * @param documentUrl The source URL to the input document.
      * @param analyzeDocumentOptions The additional configurable {@link AnalyzeDocumentOptions options} that may be
      * passed when analyzing documents.
@@ -88,7 +128,7 @@ public final class DocumentAnalysisClient {
      *
      * @return A {@link SyncPoller} to poll the progress of the analyze document operation until it has completed,
      * has failed, or has been cancelled. The completed operation returns an {@link AnalyzeResult}.
-     * @throws DocumentAnalysisException If analyze operation fails and the {@link AnalyzeResultOperation}returns
+     * @throws DocumentModelOperationException If analyze operation fails and the {@link AnalyzeResultOperation} returns
      * with an {@link OperationStatus#FAILED}.
      * @throws IllegalArgumentException If {@code documentUrl} or {@code modelId} is null.
      */
@@ -107,16 +147,34 @@ public final class DocumentAnalysisClient {
      * error message indicating absence of cancellation support.</p>
      *
      * <p><strong>Code sample</strong></p>
-     * {@codesnippet com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long}
+     * <!-- src_embed com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long -->
+     * <pre>
+     * File document = new File&#40;&quot;&#123;local&#47;file_path&#47;fileName.jpg&#125;&quot;&#41;;
+     * String modelId = &quot;&#123;custom_trained_model_id&#125;&quot;;
+     * byte[] fileContent = Files.readAllBytes&#40;document.toPath&#40;&#41;&#41;;
+     * try &#40;InputStream targetStream = new ByteArrayInputStream&#40;fileContent&#41;&#41; &#123;
      *
-     * @param modelId The unique model ID to be used or the supported prebuilt models - "prebuilt-receipt",
-     * "prebuilt-businessCard", "prebuilt-idDocument", "prebuilt-document", "prebuilt-invoice", "prebuilt-layout".
+     *     documentAnalysisClient.beginAnalyzeDocument&#40;modelId, targetStream, document.length&#40;&#41;&#41;
+     *         .getFinalResult&#40;&#41;
+     *         .getDocuments&#40;&#41;.stream&#40;&#41;
+     *         .map&#40;AnalyzedDocument::getFields&#41;
+     *         .forEach&#40;documentFieldMap -&gt; documentFieldMap.forEach&#40;&#40;key, documentField&#41; -&gt; &#123;
+     *             System.out.printf&#40;&quot;Field text: %s%n&quot;, key&#41;;
+     *             System.out.printf&#40;&quot;Field value data content: %s%n&quot;, documentField.getContent&#40;&#41;&#41;;
+     *             System.out.printf&#40;&quot;Confidence score: %.2f%n&quot;, documentField.getConfidence&#40;&#41;&#41;;
+     *         &#125;&#41;&#41;;
+     * &#125;
+     * </pre>
+     * <!-- end com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long -->
+     *
+     * @param modelId The unique model ID to be used. Use this to specify the custom model ID or prebuilt model ID.
+     * Prebuilt model IDs supported can be found <a href="https://aka.ms/azsdk/formrecognizer/models">here</a>
      * @param document The data of the document to analyze information from.
      * @param length The exact length of the data.
      *
      * @return A {@link SyncPoller} that polls the of progress of analyze document operation until it has completed,
      * has failed, or has been cancelled. The completed operation returns an {@link AnalyzeResult}.
-     * @throws DocumentAnalysisException If analyze operation fails and the {@link AnalyzeResultOperation}returns
+     * @throws DocumentModelOperationException If analyze operation fails and the {@link AnalyzeResultOperation}returns
      * with an {@link OperationStatus#FAILED}.
      * @throws IllegalArgumentException If {@code document} or {@code modelId} is null.
      */
@@ -133,10 +191,30 @@ public final class DocumentAnalysisClient {
      * error message indicating absence of cancellation support.</p>
      *
      * <p><strong>Code sample</strong></p>
-     * {@codesnippet com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long-AnalyzeDocumentOptions-Context}
+     * <p> Analyze a document with configurable options. </p>
+     * <!-- src_embed com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long-AnalyzeDocumentOptions-Context -->
+     * <pre>
+     * File document = new File&#40;&quot;&#123;local&#47;file_path&#47;fileName.jpg&#125;&quot;&#41;;
+     * String modelId = &quot;&#123;custom_trained_model_id&#125;&quot;;
+     * byte[] fileContent = Files.readAllBytes&#40;document.toPath&#40;&#41;&#41;;
      *
-     * @param modelId The unique model ID to be used or the supported prebuilt models - "prebuilt-receipt",
-     * "prebuilt-businessCard", "prebuilt-idDocument", "prebuilt-document", "prebuilt-invoice", "prebuilt-layout".
+     * try &#40;InputStream targetStream = new ByteArrayInputStream&#40;fileContent&#41;&#41; &#123;
+     *     documentAnalysisClient.beginAnalyzeDocument&#40;modelId, targetStream, document.length&#40;&#41;,
+     *             new AnalyzeDocumentOptions&#40;&#41;.setPages&#40;Arrays.asList&#40;&quot;1&quot;, &quot;3&quot;&#41;&#41;, Context.NONE&#41;
+     *         .getFinalResult&#40;&#41;
+     *         .getDocuments&#40;&#41;.stream&#40;&#41;
+     *         .map&#40;AnalyzedDocument::getFields&#41;
+     *         .forEach&#40;documentFieldMap -&gt; documentFieldMap.forEach&#40;&#40;key, documentField&#41; -&gt; &#123;
+     *             System.out.printf&#40;&quot;Field text: %s%n&quot;, key&#41;;
+     *             System.out.printf&#40;&quot;Field value data content: %s%n&quot;, documentField.getContent&#40;&#41;&#41;;
+     *             System.out.printf&#40;&quot;Confidence score: %.2f%n&quot;, documentField.getConfidence&#40;&#41;&#41;;
+     *         &#125;&#41;&#41;;
+     * &#125;
+     * </pre>
+     * <!-- end com.azure.ai.formrecognizer.DocumentAnalysisClient.beginAnalyzeDocument#string-InputStream-long-AnalyzeDocumentOptions-Context -->
+     *
+     * @param modelId The unique model ID to be used. Use this to specify the custom model ID or prebuilt model ID.
+     * Prebuilt model IDs supported can be found <a href="https://aka.ms/azsdk/formrecognizer/models">here</a>
      * @param document The data of the document to analyze information from.
      * @param length The exact length of the data.
      * @param analyzeDocumentOptions The additional configurable {@link AnalyzeDocumentOptions options} that may be
@@ -145,7 +223,7 @@ public final class DocumentAnalysisClient {
      *
      * @return A {@link SyncPoller} that polls the of progress of analyze document operation until it has completed,
      * has failed, or has been cancelled. The completed operation returns an {@link AnalyzeResult}.
-     * @throws DocumentAnalysisException If analyze operation fails and the {@link AnalyzeResultOperation} returns
+     * @throws DocumentModelOperationException If analyze operation fails and the {@link AnalyzeResultOperation} returns
      * with an {@link OperationStatus#FAILED}.
      * @throws IllegalArgumentException If {@code document} or {@code modelId} is null.
      */