+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/jms/AbstractServiceBusJmsAutoConfiguration.java
@@ -70,9 +70,6 @@ public abstract class AbstractServiceBusJmsAutoConfiguration {
 
     private void configureTopicListenerContainerFactory(DefaultJmsListenerContainerFactory jmsListenerContainerFactory) {
         AzureServiceBusJmsProperties.Listener listener = azureServiceBusJMSProperties.getListener();
-        if (azureServiceBusJMSProperties.getTopicClientId() != null) {
-            jmsListenerContainerFactory.setClientId(azureServiceBusJMSProperties.getTopicClientId());
-        }
         if (listener.isReplyPubSubDomain() != null) {
             jmsListenerContainerFactory.setReplyPubSubDomain(listener.isReplyPubSubDomain());
         }