+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/jms/NonPremiumServiceBusJmsAutoConfiguration.java
@@ -6,6 +6,7 @@ package com.azure.spring.cloud.autoconfigure.jms;
 import com.azure.spring.cloud.autoconfigure.jms.properties.AzureServiceBusJmsProperties;
 import com.azure.spring.core.connectionstring.implementation.ServiceBusConnectionString;
 import org.apache.qpid.jms.JmsConnectionFactory;
+import org.apache.qpid.jms.policy.JmsDefaultPrefetchPolicy;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -50,6 +51,15 @@ public class NonPremiumServiceBusJmsAutoConfiguration extends AbstractServiceBus
         jmsConnectionFactory.setClientID(clientId);
         jmsConnectionFactory.setUsername(sasKeyName);
         jmsConnectionFactory.setPassword(sasKey);
+
+        JmsDefaultPrefetchPolicy prefetchPolicy = (JmsDefaultPrefetchPolicy) jmsConnectionFactory.getPrefetchPolicy();
+        AzureServiceBusJmsProperties.PrefetchPolicy prefetchProperties = serviceBusJMSProperties.getPrefetchPolicy();
+        prefetchPolicy.setDurableTopicPrefetch(prefetchProperties.getDurableTopicPrefetch());
+        prefetchPolicy.setQueueBrowserPrefetch(prefetchProperties.getQueueBrowserPrefetch());
+        prefetchPolicy.setQueuePrefetch(prefetchProperties.getQueuePrefetch());
+        prefetchPolicy.setTopicPrefetch(prefetchProperties.getTopicPrefetch());
+        jmsConnectionFactory.setPrefetchPolicy(prefetchPolicy);
+
         return jmsConnectionFactory;
     }